separated dto from model
