+++ src/test/java/com/bank/moneytransfer/controller/TransferAndGetIntegrationParallelTest.java
@@ -4,9 +4,9 @@ import com.anarsoft.vmlens.concurrent.junit.ConcurrentTestRunner;
 import com.bank.moneytransfer.TestUtil;
 import com.bank.moneytransfer.datastore.BankAccountStorage;
 import com.bank.moneytransfer.exception.ExceptionHandler;
-import com.bank.moneytransfer.model.AllBankAccountsResponse;
+import com.bank.moneytransfer.dto.AllBankAccountsResponse;
 import com.bank.moneytransfer.model.BankAccount;
-import com.bank.moneytransfer.model.TransferMoneyRequest;
+import com.bank.moneytransfer.dto.TransferMoneyRequest;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
 import org.junit.After;
