+++ src/test/java/com/bank/moneytransfer/controller/TransferMoneyFunctionalTest.java
@@ -5,8 +5,8 @@ import com.bank.moneytransfer.datastore.BankAccountStorage;
 import com.bank.moneytransfer.exception.ErrorMessages;
 import com.bank.moneytransfer.exception.ExceptionHandler;
 import com.bank.moneytransfer.model.BankAccount;
-import com.bank.moneytransfer.model.TransferMoneyRequest;
-import com.bank.moneytransfer.model.TransferMoneyResponse;
+import com.bank.moneytransfer.dto.TransferMoneyRequest;
+import com.bank.moneytransfer.dto.TransferMoneyResponse;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
 import org.junit.BeforeClass;
