+++ src/test/java/com/bank/moneytransfer/service/BankAccountsServiceTest.java
@@ -3,8 +3,8 @@ package com.bank.moneytransfer.service;
 import com.bank.moneytransfer.exception.ErrorMessages;
 import com.bank.moneytransfer.exception.types.AccountPresentException;
 import com.bank.moneytransfer.exception.types.IllegalAccountParamsException;
-import com.bank.moneytransfer.model.AddBankAccountRequest;
-import com.bank.moneytransfer.model.AllBankAccountsResponse;
+import com.bank.moneytransfer.dto.AddBankAccountRequest;
+import com.bank.moneytransfer.dto.AllBankAccountsResponse;
 import com.bank.moneytransfer.model.BankAccount;
 import org.junit.Before;
 import org.junit.Test;
